+++ gazebo/rendering/Grid.cc
@@ -273,21 +273,9 @@ void Grid::Enable(const bool _enable)
 }
 
 //////////////////////////////////////////////////
-Ogre::SceneNode *Grid::GetSceneNode()
-{
-  return this->dataPtr->gridVis->GetSceneNode();
-}
-
-//////////////////////////////////////////////////
 VisualPtr Grid::GridVisual() const
 {
   return this->dataPtr->gridVis;
-}
-
-//////////////////////////////////////////////////
-common::Color Grid::GetColor() const
-{
-  return this->Color();
 }
 
 //////////////////////////////////////////////////
@@ -297,21 +285,9 @@ common::Color Grid::Color() const
 }
 
 //////////////////////////////////////////////////
-uint32_t Grid::GetCellCount() const
-{
-  return this->CellCount();
-}
-
-//////////////////////////////////////////////////
 uint32_t Grid::CellCount() const
 {
   return this->dataPtr->cellCount;
-}
-
-//////////////////////////////////////////////////
-float Grid::GetCellLength() const
-{
-  return this->CellLength();
 }
 
 //////////////////////////////////////////////////
@@ -321,22 +297,10 @@ float Grid::CellLength() const
 }
 
 //////////////////////////////////////////////////
-float Grid::GetLineWidth() const
-{
-  return this->LineWidth();
-}
-
-//////////////////////////////////////////////////
 float Grid::LineWidth() const
 {
   gzwarn << "Line width is currently not supported. Issue #1978" << std::endl;
   return this->dataPtr->lineWidth;
-}
-
-//////////////////////////////////////////////////
-uint32_t Grid::GetHeight() const
-{
-  return this->Height();
 }
 
 //////////////////////////////////////////////////