+++ gazebo/rendering/MovableText.hh
@@ -25,7 +25,6 @@
 #include "gazebo/rendering/ogre_gazebo.h"
 
 #include "gazebo/common/Color.hh"
-#include "gazebo/math/Box.hh"
 #include "gazebo/util/system.hh"
 
 namespace boost
@@ -146,12 +145,6 @@ namespace gazebo
 
       /// \brief Get the axis aligned bounding box of the text.
       /// \return The axis aligned bounding box.
-      /// \deprecated See ignition::math::Box AABB();
-      public: math::Box GetAABB() GAZEBO_DEPRECATED(8.0);
-
-      /// \brief Get the axis aligned bounding box of the text.
-      /// \return The axis aligned bounding box.
-      /// \deprecated See ignition::math::Box AABB();
       public: ignition::math::Box AABB();
 
       /// \brief Update the text.